Use extension methods to register services
